+++ src/Schemas/ColumnSchemaBuilder.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Schemas;
 
-use Yiisoft\Db\Drivers\Connection;
+use Yiisoft\Db\Connection\Connection;
 use Yiisoft\Db\Expressions\Expression;
 use Yiisoft\Strings\StringHelper;
